Move List widget into own file
